Added javadoc for new files/packages
