+++ server/src/main/java/org/opensearch/common/cache/tier/OnHeapCachingTier.java
@@ -6,13 +6,11 @@
  * compatible open source license.
  */
 
-package org.opensearch.indices;
+package org.opensearch.common.cache.tier;
 
 /**
- * This is specific to disk caching tier and can be used to add methods which are specific to disk tier.
+ * This is specific to onHeap caching tier and can be used to add methods which are specific to this tier.
  * @param <K> Type of key
  * @param <V> Type of value
  */
-public interface DiskCachingTier<K, V> extends CachingTier<K, V> {
-
-}
+public interface OnHeapCachingTier<K, V> extends CachingTier<K, V> {}
