+++ server/src/main/java/org/opensearch/common/cache/tier/TieredCacheEventListener.java
@@ -6,10 +6,15 @@
  * compatible open source license.
  */
 
-package org.opensearch.indices;
+package org.opensearch.common.cache.tier;
 
 import org.opensearch.common.cache.RemovalNotification;
 
+/**
+ * This can be used to listen to tiered caching events
+ * @param <K> Type of key
+ * @param <V> Type of value
+ */
 public interface TieredCacheEventListener<K, V> {
 
     void onMiss(K key, TierType tierType);
